+++ tiny-lib/Utils.cpp
@@ -6,7 +6,7 @@ std::string Utils::ByteArrayToHexString(const std::vector<uint8_t>& vec)
     std::string hash;
     hash.resize(vec.size() * 2);
 
-    for (int i = 0; i < vec.size(); i++)
+    for (size_t i = 0; i < vec.size(); i++)
         sprintf_s(hash.data() + (i * 2), 2, "%02x", vec[i]);
 
     return hash;
